fix cpu utilization calculation
retry count during retry calibration can overflow, prevent overflow by
using uint64_t instead of int.

+++ src/common.h
@@ -79,7 +79,7 @@ struct wq_info {
 
 struct poll_cnt {
 	union {
-		int retry;
+		uint64_t retry;
 		int monitor;
 	};
 	int mwait;
